correcion 2 de entrega 1

+++ models/apis/carrito.api.js
@@ -12,7 +12,7 @@ class CarritoAPI {
     getProductsById = async id => {
         const carritos = await this.getAll();
         const foundCart = carritos.find(c => c.id == id);
-        return !foundCart ? { error: 'Carrito no encontrado' } : foundCart.products;
+        return !foundCart ? { error: 'Carrito no encontrado' } : foundCart.productos;
     }
     getAll = async () => {
         try {
@@ -26,17 +26,30 @@ class CarritoAPI {
         carritos.push({ id: newId, ...cart });
         try {
             await fs.writeFile(this.path, JSON.stringify(carritos, null, 4));
-            return { id: newId, ...cart };
+            return { caritoID: newId };
         } catch (error) { throw new Error('error al crear carrito') }
     }
-    addProductsToCart = async (id, products) => {
+    addProductsToCart = async (id, id_product, quantity) => {
         const carritos = await this.getAll();
         const foundIndex = carritos.findIndex(c => c.id == id);
         if (!carritos[foundIndex]) return { error: 'carrito inexistente' };
-        carritos[foundIndex].products = products;
+
+        let parsedProducts;
+        try {
+            const productos = await fs.readFile('./data/productos.json');
+            parsedProducts = JSON.parse(productos);
+        } catch (error) {
+            return { error: 'no se pudo obtener los productos' }
+        }
+
+        const indexProduct = parsedProducts.findIndex(p => p.id == id_product);
+        if (!parsedProducts[indexProduct]) return { error: 'id de producto inexistente' };
+
+        for (let i = 0; i < quantity; i++) { carritos[foundIndex].productos.push(parsedProducts[indexProduct]) }
+
         try {
             await fs.writeFile(this.path, JSON.stringify(carritos, null, 4));
-            return carritos[foundIndex];
+            return carritos[foundIndex].productos;
         } catch (error) { throw new Error('no se pudo guardar los productos en el carrito') }
     }
     deleteCart = async id => {
@@ -53,9 +66,9 @@ class CarritoAPI {
         const carritos = await this.getAll();
         const indexCart = carritos.findIndex(c => c.id == id);
         if (!carritos[indexCart]) return { error: 'Carrito inexistente' };
-        const indexProduct = carritos[indexCart].products.findIndex(p => p.id == id_prod);
-        if (!carritos[indexCart].products[indexProduct]) return { error: 'Producto inexistente en carrito' };
-        const deletedProduct = carritos[indexCart].products.splice(indexProduct, 1);
+        const indexProduct = carritos[indexCart].productos.findIndex(p => p.id == id_prod);
+        if (!carritos[indexCart].productos[indexProduct]) return { error: 'Producto inexistente en carrito' };
+        const deletedProduct = carritos[indexCart].productos.splice(indexProduct, 1);
         try {
             await fs.writeFile(this.path, JSON.stringify(carritos, null, 4));
             return deletedProduct;
